Fixing output paths for glob supplementals
This requires copy-webpack-plugin 4.6.0 for the transformPath property, which
also requires working around a yet-unfixed regression in c-w-p that was causing
a test failure - the "nobraces" option no longer works properly when the input
is an explicit file or directory. So, if there's a cwpPatternConfig.fromArgs
option, we explicitly tell c-w-p to treat this as a glob. We could probably
safely do this no matter what, but better safe than sorry.